requirements.txt
fastapi
uvicorn[standard]
requests
pillow
numpy
PyYAML

# Note: torch, torchvision, and CUDA are included in the PyTorch base image
# SAM3 will be installed from source in the Dockerfile
